requirements-cpu.txt
matplotlib
datasets
torch
torchvision
wandb
numpy<2
pillow